[SC-147] Outsource fitting into new class HPIFitData
