+++ MNE/inverse/inverse.pro
@@ -113,7 +113,8 @@ SOURCES += \
     c/mne_inverse_operator.cpp \
     c/mne_meas_data.cpp \
     c/mne_meas_data_set.cpp \
-    hpiFit/hpifit.cpp
+    hpiFit/hpifit.cpp \
+    hpiFit/hpifitdata.cpp
 
 
 HEADERS +=\
@@ -134,7 +135,8 @@ HEADERS +=\
     c/mne_inverse_operator.h \
     c/mne_meas_data.h \
     c/mne_meas_data_set.h \
-    hpiFit/hpifit.h
+    hpiFit/hpifit.h \
+    hpiFit/hpifitdata.h
 
 
 INCLUDEPATH += $${EIGEN_INCLUDE_DIR}
