+++ applications/mne_rt_server/connectors/FiffSimulator/fiffsimulator.cpp
@@ -108,8 +108,7 @@ const QString FiffSimulator::Commands::SIMFILE      = "simfile";
 
 FiffSimulator::FiffSimulator()
 : m_pFiffProducer(new FiffProducer(this))
-//, m_sResourceDataPath(QString("./MNE-sample-data/MEG/sample/sample_audvis_raw.fif").arg(QCoreApplication::applicationDirPath()))
-, m_sResourceDataPath(QString("D:/CloudStation/Dokumente/Karriere/TU Ilmenau/Promotion/Projekte/2015_BabyMEG/HPI fitting/2017_02_24 real_baby_measurement/170224_100033_4872515_Run2_HPI_0.1_raw.fif").arg(QCoreApplication::applicationDirPath()))
+, m_sResourceDataPath(QString("./MNE-sample-data/MEG/sample/sample_audvis_raw.fif").arg(QCoreApplication::applicationDirPath()))
 , m_uiBufferSampleSize(100)//(4)
 , m_AccelerationFactor(1.0)
 , m_TrueSamplingRate(0.0)
